requirements.txt
requests
prometheus-client
falcon >=3
PyYAML
argparse
pyeapi
